Add roll counters to Need/Greed frames
Added:
- Roll counters for group mates on N/G/P buttons for each loot window.
- Track all items you've rolled on in a blacklist, so they don't "pop
back" for a second after you roll (occasional bug)

Issues:
- Since we're only passed an Item when a roll occurs (and it has no
LootId), it's impossible to know which of two duplicate items the user
might have wanted to roll on. Right now, their rolls shows on both
frames - so one user hitting Need on one item and Pass on the other
would show 1 Need and 1 Pass on both of those duplicate items to the
player. There are alternatives to this (including grouping duplicate
items into one roll frame), but I think this may be the best way to
display this info.

I believe this is also only true for 'basic' items, not actual loot
people would more likely be rolling on.