Fix Nameplate Width and Height labels
Fixed the text descriptions in the Nameplate options pane.

Note: The underlying code still incorrectly refers to the fields by the
opposite name (nBarWidth for the actual height, and vice versa). I've
left this as-is to not cause issues for those with saved height/width
values. This could be fixed by adding a dummy 'bWidthHeightFixed'
variable set to false by default, and a small function to swap width and
height values and then set that boolean to true.